+++ src/KyoshinEewViewer/Series/KyoshinMonitor/ReplayFileEarthquakeInformationHost.cs
@@ -109,7 +109,7 @@ public partial class ReplayFileEarthquakeInformationHost : EarthquakeInformation
 					{
 						{
 							LandLayerType.EarthquakeInformationSubdivisionArea,
-							intensityAreas.ToDictionary(p => p.Key, p => FixedObjectRenderer.IntensityPaintCache[p.Value].Background.Color)
+							intensityAreas.ToDictionary(p => (long)p.Key, p => FixedObjectRenderer.IntensityPaintCache[p.Value].Background.Color)
 						},
 					}
 				};
@@ -123,7 +123,7 @@ public partial class ReplayFileEarthquakeInformationHost : EarthquakeInformation
 					{
 						{
 							LandLayerType.EarthquakeInformationSubdivisionArea,
-							warningAreaCodes.ToDictionary(c => c, c => SKColors.Tomato)
+							warningAreaCodes.ToDictionary(c => (long)c, c => SKColors.Tomato)
 						},
 					}
 				};
